修改精准报价相关接口，添加删除接口

+++ database/src/main/java/cn/thinkfree/database/pcvo/QuotationVo.java
@@ -11,6 +11,7 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @ApiModel("精准报价实体")
@@ -23,13 +24,13 @@ public class QuotationVo {
     @ApiModelProperty(value = "基础施工项实体")
     private List<BasisConstructionVO> BasicsTableData;
     @ApiModelProperty(value = "基础施工项总价")
-    private Integer basicsSumPrice;
+    private BigDecimal basicsSumPrice;
     @ApiModelProperty(value = "软装实体")
     private List<SoftQuoteVO> MaterialTableData;
     @ApiModelProperty(value = "软装总价")
-    private Integer materialSumPrice;
+    private BigDecimal materialSumPrice;
     @ApiModelProperty(value = "硬装实体")
     private List<HardQuoteVO> furnitureTableData;
     @ApiModelProperty(value = "硬装总价")
-    private Integer furnitureSumPrice;
+    private BigDecimal furnitureSumPrice;
 }
